Use 64 bits unsigned for significant sb_map

+++ libovvc/vcl_transform_unit.c
@@ -12,7 +12,7 @@
 
 struct TBInfo {
    uint16_t last_pos;
-   uint32_t sig_sb_map;
+   uint64_t sig_sb_map;
 };
 
 struct TUInfo {
